style: (#164) 패키지 이름 변경
