+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/auth/mapper/AuthCodeMapper.kt
@@ -3,7 +3,7 @@ package team.aliens.dms.persistence.auth.mapper
 import org.springframework.stereotype.Component
 import team.aliens.dms.domain.auth.model.AuthCode
 import team.aliens.dms.persistence.GenericMapper
-import team.aliens.dms.persistence.auth.model.AuthCodeEntity
+import team.aliens.dms.persistence.auth.entity.AuthCodeEntity
 
 @Component
 class AuthCodeMapper : GenericMapper<AuthCode, AuthCodeEntity> {
