+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/auth/repository/AuthCodeEntityRepository.kt
@@ -3,7 +3,7 @@ package team.aliens.dms.persistence.auth.repository
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 import team.aliens.dms.domain.auth.model.EmailType
-import team.aliens.dms.persistence.auth.model.AuthCodeEntity
+import team.aliens.dms.persistence.auth.entity.AuthCodeEntity
 
 @Repository
 interface AuthCodeEntityRepository : CrudRepository<AuthCodeEntity, String> {
