+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/auth/repository/AuthCodeLimitRepository.kt
@@ -3,7 +3,7 @@ package team.aliens.dms.persistence.auth.repository
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 import team.aliens.dms.domain.auth.model.EmailType
-import team.aliens.dms.persistence.auth.model.AuthCodeLimitEntity
+import team.aliens.dms.persistence.auth.entity.AuthCodeLimitEntity
 import java.util.UUID
 
 @Repository
