+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/auth/repository/RefreshTokenRepository.kt
@@ -2,7 +2,7 @@ package team.aliens.dms.persistence.auth.repository
 
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
-import team.aliens.dms.persistence.auth.model.RefreshTokenEntity
+import team.aliens.dms.persistence.auth.entity.RefreshTokenEntity
 
 @Repository
 interface RefreshTokenRepository : CrudRepository<RefreshTokenEntity, String> {
